requirements.txt
kafka-python
pyspark
flask